backend/requirements.txt
flask
flask-cors
tensorflow>=2.12.0
pillow
numpy>=1.23.5
scikit-learn>=1.0.2
matplotlib>=3.5.3
pandas>=1.5.3 